server/src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.121
botocore==1.34.121
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.21
django-nested-admin==4.0.2
django-nine==0.2.7
django-storages==1.14.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
elastic-transport==8.13.1
elasticsearch==8.0.0
elasticsearch-dsl==8.9.0
gunicorn==23.0.0
flower==2.0.1
humanize==4.9.0
idna==3.7
jmespath==1.0.1
kombu==5.3.7
packaging==24.0
pathvalidate==3.2.0
pillow==10.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.46
psycopg2==2.9.9
pyClamd==0.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-monkey-business==1.0.0
pytz==2024.1
redis==5.0.4
requests==2.32.2
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tornado==6.4
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
